requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.7
django-filter==2.1.0
djangorestframework==3.9.1
idna==2.8
Markdown==3.0.1
Pillow==5.4.1
pytz==2018.9
requests==2.21.0
urllib3==1.24.1
virtualenv==16.4.1
virtualenvwrapper-win==1.2.5
